Comparator5Bit
This device is customized for the forwording control logic, since it compares just 5 bits of the instruction register. The already designed comparators, the former for BTB, the latter in the ALU, are not much feasible. In fact, one is optimized for power of 2s comparison operands, the other else computes and takes care lt and gt scenarios